--- src/web/routes/application/steps/confirm/index.js
@@ -1,5 +1,0 @@
-const { registerConfirmRoute } = require('./confirm')
-
-module.exports = {
-  registerConfirmRoute
-}
