+++ src/web/routes/application/steps/decision/index.js
@@ -1,0 +1,5 @@
+const { registerConfirmRoute } = require('./decision')
+
+module.exports = {
+  registerConfirmRoute
+}
